+++ applications/network/lib/middleware/common/logger/index.ts
