+++ applications/network/lib/middleware/common/logger/logger.token.ts
