+++ applications/network/lib/middleware/index.ts
@@ -1,4 +1,2 @@
-export { default as pingPong } from "./ping-pong.middleware";
-export { default as version } from "./version.middleware";
-
-export * from "./logger";
+export * from "./common";
+export * from "./service";
