+++ applications/network/lib/middleware/service/index.ts
@@ -1,0 +1,2 @@
+export { default as pingPong } from "./ping-pong.middleware";
+export { default as version } from "./version.middleware";
