+++ applications/network/lib/middleware/service/ping-pong.middleware.ts
@@ -1,6 +1,6 @@
 import Application from "koa";
-import Context from "../context";
-import State from "../state";
+import Context from "../../context";
+import State from "../../state";
 
 function pingPong(): Application.Middleware<State, Context> {
   return async (context, next) => {
